pyproject.toml
[tool.poetry]
name = "luhack_bot"
version = "0.1.0"
description = "LUHack's discord bot"
authors = ["Ben Simms <ben@bensimms.moe>"]

[tool.poetry.dependencies]
python = ">=3.8,<4"
itsdangerous = "^2.0"
cryptography = "^37.0"
gino = "^1.0"
python-dotenv = "^0.21"
sqlalchemy_utils = "^0.38"
asyncpg = "^0.26"
alembic = "^1.4"
starlette = "^0.20"
jinja2 = "^3.0.1"
python-multipart = "^0.0"
pygments = "^2.6"
aiofiles = "^22.1"
slug = "^2.0"
WTForms = "^3.0"
yarl = "^1.6"
gunicorn = "^20.0"
PyCrypto = "^2.6"
python-dateutil = "^2.8"
parsedatetime = "^2.6"
"discord.py" = "^2.0"
uvicorn = "^0.18"
psycopg2-binary = "^2.9"
tabulate = "^0.8"
aiosmtplib = "^1.1.4"
mistune = "2.0.4"
orjson = "^3.6"
Jinja2 = "^3.0.1"
Authlib = "^1.0.1"
SQLAlchemy-Searchable = "^1.4.1"
httpx = "^0.23.0"

[tool.poetry.dev-dependencies]
psycopg2-binary = "^2.8"
pyre-check = "^0.9.1"

[tool.poetry.scripts]
start_bot = 'luhack_bot:run'
gen_tokens = 'luhack_bot:gen_tokens'
export_content = 'luhack_bot:export_writeups'

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.core.masonry.api"
